requirements.txt
beautifulsoup4==4.5.1
dj-database-url==0.4.1
Django==1.10.3
djangorestframework==3.5.3
gevent==1.1.2
psycogreen==1.0
psycopg2==2.6.2
requests==2.12.1
uWSGI==2.0.13.1
